Don't need to modify activation dimensions.
When passing `Z` and `A` to backprop. We need gradients for all weights,
not only the weights of the activations used to calculate the cost
(because we use a softmax activation in the final layer, which outputs a
vector with elements for each class, we only use a subset of the final
activations in the cost function - those activations that are the predictions
for the label corresponding to a particular example).