Update backprop to properly apply the chain rule.
For each layer beyond the output layer, needed to calculate the derivative of
layer `L`'s linear activation with respect to layer `L-1`'s activation (output)
in order to then calculate the derivative of `L-1`'s activation with
respect to its weights and biases.

It is this multiplication of one layer's linear activation by the
previous layer's activation that propagates the gradients backward
through the network.