Convert labels (`Y`) to a "one-hot" matrix.
This is essentially converting categorical data, specifically the labels
saying which class each example belongs to, from integers in the range
`0..number_of_classes` to vectors with `number_of_classes` dimensions
where every element is either 0 or 1. A 1 at the `ith` position in the
one-hot vector denotes that the example is of class `i`.