Remove unnecessary thresholding of values.
Originally I added code to ensure that values in forward prop and the
cost function didn't overflow or underflow, but this turned out to be
unnecessary. The reaason this numerical instability was occuring was
because I wasn't computing the analytic gradients correctly during
backprop. This resulted in gradients that were far too large, which in
turn caused we network's weights to become too large, which finally
resulted in super high values in forward prop and the cost function as
more and more iterations were performed.