Smaller learning rate & better logging of cost.
The learning rate of `0.025` was much too high, and would regularly
cause the optimizer to overshoot the minimum after it updated the
parameters.

The Adam paper recommends a learning rate of `0.001`, and rates of
approximately that magnitude seem to overshoot much less (the goal is to
not overshoot at all, and instead decrease the cost monotonically).

The following changes should help make it obvious when the learning rate
is too high:

+ Print the latest cost to the console every 10 iterations instead of
every 100. This makes it harder to miss overshooting.

+ Lower the learning rate by an order of magnitude. This means we will
decrease cost more slowly on any one given update, but will decrease it
far more consistently, resulting in a lower cost in the end.

+ Print a very noticeable message whenever the cost increases.