Properly normalize data.
+ Use proper axis argument when finding mean and standard deviation for
normalization. Previously the normalization code was calculating the mean and
standard deviation for each *example*, not each *feature* as it
should've been.

+ Apply normalization to the test set based on the mean and standard
deviation of the training set. This is based on the assumption that the
training data is representative of the test data and any data the
algorithm will see "in the wild". If that assumption holds true then we
can apply normalization to each new example the algorithm sees.

+ And a better variable name: `d` -> `data`.