Try training a 7-class classifier.
The FER dataset provides 7 classes. Originally we were coverting those 7
classes into the 3 facial expressions we want the web interface to be
able to animate to: neutral, smiling, and frowning. This meant adding an
implicit conversion step at the machine learning layer that may have
been affecting the performance of the classifier on the test set.

In this new approach we train the network to classify the original 7 classes,
and have the web interface handle choosing the correct expression to
animate into given one of those 7 classes.

Even if this doesn't fix the algorithm's performance on the test set
this is a better way to go because it simplifies the code greatly.